Выполнено задание 5 раздела. mian.js разбит на модули.
